requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
black==24.1.1
click==8.1.7
Django==4.2.10
djangorestframework==3.14.0
Faker==23.2.0
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.3
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uuid==1.30